feat: add `[grind ←=]` attribute
This PR adds support for equality backward reasoning to `grind`.
We can illustrate the new feature with the following example. Suppose we have a theorem:
```lean
theorem inv_eq {a b : α} (w : a * b = 1) : inv a = b
```
and we want to instantiate the theorem whenever we are tying to prove
`inv t = s` for some terms `t` and `s`
The attribute `[grind ←]` is not applicable in this case because, by default,
`=` is not eligible for E-matching. The new attribute `[grind ←=]` instructs
`grind` to use the equality and consider disequalities in the `grind`
proof state as candidates for E-matching.