feat: make coeFun delaborator respect `pp.tagAppFns`
This PR makes the pretty printer for `.coeFun`-tagged functions respect how constants are annotated when `pp.tagAppFns` is true. The effect is that in docgen, when an expression pretty prints as `f x y z`, where this is implicitly a function coercion, then if `f` is a constant it will be linkified.