perf: fast path for multiplication with constants in bv_decide (#6739)
This PR adds a fast path for bitblasting multiplication with constants
in `bv_decide`.

While the circuit generated is the same (as the AIG already performs
constant folding) this avoids calling out to the shift and addition
bitblaster unless required. Thus the overall time to generate the
circuit is reduced. Inspired by
[bitwuzla](https://github.com/bitwuzla/bitwuzla/blob/25d77f819c84575e6b759db0bbc4ffb2fbb4f20c/src/lib/bitblast/bitblaster.h#L454).