fix: Funind: no do unfold auxillary defs
This PR changes internals in the funidn construction.

Previously we would unfold defs `Tree.revrev._unary.proof_3` when
transforming the function body. This is somewhat expensive (lots of
proofs copied) and it may expose recursive calls within proofs that we
probably do not care about, and that can look different than what the
user wrote somewhere (`simpMatch`, looking at you here).

So we try to leave them in place now.

Also previously we did not include the “old” IH in the local context, so
that creating a MVar would not pick it up. But this prevented us from
inferring types, which I really need here.

So now let's keep them in the local context, but wrap the creation of
the metavariables in `withErasedFVars`.