doc: mention `Prop`s are equal to `True` or `False`
This PR modifies the `Prop` docstring to point out that every proposition is propositionally equal to either `True` or `False. This will help point users toward seeing that `Prop` is like `Bool`.

I considered mentioning `Classical.propComplete`, but it's probably better not making it seem like that's how you should work with propositions.