feat: make structure parent projections nameable
This PR modifies the `structure` syntax so that parents can be named, like in
```
structure S extends toParent : P
```
The syntax is also modified so that the resultant type comes *before* the `extends` clause (**breaking change**). This is necessary to prevent a parsing ambiguity. Implements RFC #7099.

Will need followup PRs for cleanup after a stage0 update.