feat: documentation links and examples
Docstrings are now pre-processed as follows:

 * Output included as part of examples is shown with leading line
 comment indicators in hovers

 * URLs of the form `lean-manual://section/section-id` are rewritten
 to links that point at the corresponding section in the Lean
 reference manual. The reference manual's base URL is configured when
 Lean is built and can be overridden with the `LEAN_MANUAL_ROOT`
 environment variable. This way, releases can point documentation
 links to the correct snapshot, and users can use their own, e.g. for
 offline reading.

Manual URLs in docstrings are validated when the docstring is
added. The presence of a URL starting with `lean-manual://` that is
not a syntactically valid section link causes the docstring to be
rejected. This allows for future extensibility to the set of allowed
links. There is no validation that the linked-to section actually
exists.

A stage0 update will be required to make the documentation site
configurable at build time and for releases.