feat: show declaration keywords in hovers
Declaration keywords are now shown in hovers as part of a
signature. This is in support of an upcoming documentation style guide
that will recommend that the text of docstrings not state what sort of
declaration something is, but instead leave it to Lean to tell
every user in a consistent manner.