feat: let synth order checker detect generalized projections
This PR modifies the instance synth order checker to detect projections-like instances, rather having special support only for constructor field projections. An instance is projection-like if the first instance implicit argument has a type that's a constant applied to all the preceding arguments in order — these arguments are assumed to be the parameters of the type.

In the future the `class` command will make all its generated parent projections use implicits for all parameters, making the generated projections all consistent.